Make MessageProgressUpdater generic
